refactor(the-rust/common-collections): restructure vectors in rust
