+++ the-rust-programming-language/common-collections/src/vector/store_multiple_types_with_enum.rs
@@ -1,0 +1,16 @@
+pub fn store_multiple_types_with_enum() {
+    #[derive(Debug)]
+    enum SpreadsheetCell {
+        Int(i32),
+        Float(f64),
+        Text(String),
+    }
+
+    let row = vec![
+        SpreadsheetCell::Int(3),
+        SpreadsheetCell::Text(String::from("blue")),
+        SpreadsheetCell::Float(10.12),
+    ];
+
+    println!("The row's element is: {row:?}");
+}
