+++ the-rust-programming-language/common-collections/src/vector/update_vector.rs
@@ -1,0 +1,11 @@
+pub fn update_value_to_vector() {
+    // Updating a Vector with `push`` method
+    let mut updated_vector = Vec::new();
+
+    updated_vector.push(5);
+    updated_vector.push(6);
+    updated_vector.push(7);
+    updated_vector.push(8);
+
+    println!("updated_vector: {updated_vector:?}");
+}
